backend/requirements.txt
fastapi==0.109.1
httpx==0.26.0
passlib==1.7.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.5
sqlalchemy==2.0.25
uvicorn==0.27.0
